requirements.txt
alembic==1.6.5
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
fastapi==0.68.0
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.1
h11==0.12.0
importlib-metadata==4.6.1
kombu==5.1.0
Mako==1.1.4
MarkupSafe==2.0.1
prompt-toolkit==3.0.19
psycopg2==2.9.1
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.22
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.15.0
vine==5.0.0
virtualenv==16.7.2
wcwidth==0.2.5
zipp==3.5.0
